Update DispatcherDefault.xml
Changes to text labels. Adding space at end and a few font or size corrections. Any other changes are not intended so if you see anything that affects you or the museum , not just my laptop, in screen position or the like that I should avoid please say so. Will be doing more text work directly in the xml file next.